Migrate to latest Kotlin, migrate to html4k 0.3.x

+++ clientside-viewimpl/src/main/kotlin/random-ticker.kt
@@ -1,11 +1,11 @@
 package market.web.impl
 
 import market.web.*
-import kotlin.js.dom.html.window
+import kotlin.browser.window
 
-fun randomInterval() = Math.max(50.0, Math.abs(Math.random() - 0.5) * 1000.0)
+fun randomInterval() = Math.max(50.0, Math.abs(Math.random() - 0.5) * 1000.0).toInt()
 
-fun startTicker(listener : QuotesListener, interval : Double = randomInterval()) {
+fun startTicker(listener : QuotesListener, interval : Int = randomInterval()) {
     window.setTimeout({
         val id = (Math.random() * 9).toInt()
         val instrument = "INSTR-$id"
